+++ app/src/main/java/com/group147/appartmentblog/screens/apartment/PostViewModel.kt
@@ -1,13 +1,26 @@
 package com.group147.appartmentblog.screens.apartment
 
+import android.util.Log
+import android.view.View
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
+import androidx.lifecycle.lifecycleScope
+import com.bumptech.glide.Glide
+import com.google.firebase.firestore.GeoPoint
+import com.group147.appartmentblog.R
+import com.group147.appartmentblog.databinding.FragmentAddApartmentBinding
+import com.group147.appartmentblog.databinding.FragmentPostBinding
 import com.group147.appartmentblog.model.Post
+import com.group147.appartmentblog.repositories.PostRepository
+import com.group147.appartmentblog.screens.apartment.PostFragment
 import com.group147.appartmentblog.util.geoToAdress.getGoogleAddressFromLatLng
+import kotlinx.coroutines.launch
+import java.util.Date
 
-class PostViewModel : ViewModel() {
+class PostViewModel(private val binding: FragmentPostBinding) : ViewModel() {
 
+    private var isEditMode = false
     private val _post = MutableLiveData<Post>()
     val post: LiveData<Post> = _post
 
@@ -27,4 +40,62 @@ class PostViewModel : ViewModel() {
 
         return address
     }
+
+
+    fun setupEditButton() {
+        binding.editButton.setOnClickListener {
+            toggleEditMode(true)
+        }
+
+        binding.saveButton.setOnClickListener {
+            if (validateInput()) {
+                updatePost()
+                toggleEditMode(false)
+            }
+        }
+    }
+
+    private fun toggleEditMode(editMode: Boolean) {
+        isEditMode = editMode
+
+        binding.titleTextView.visibility = if (editMode) View.GONE else View.VISIBLE
+        binding.contentTextView.visibility = if (editMode) View.GONE else View.VISIBLE
+        binding.priceTextView.visibility = if (editMode) View.GONE else View.VISIBLE
+        binding.roomsTextView.visibility = if (editMode) View.GONE else View.VISIBLE
+        binding.floorTextView.visibility = if (editMode) View.GONE else View.VISIBLE
+
+        binding.titleEditText.visibility = if (editMode) View.VISIBLE else View.GONE
+        binding.contentEditText.visibility = if (editMode) View.VISIBLE else View.GONE
+        binding.priceEditText.visibility = if (editMode) View.VISIBLE else View.GONE
+        binding.roomsEditText.visibility = if (editMode) View.VISIBLE else View.GONE
+        binding.floorEditText.visibility = if (editMode) View.VISIBLE else View.GONE
+
+        binding.editButton.visibility = if (editMode) View.GONE else View.VISIBLE
+        binding.addressTextView.visibility = if (editMode) View.GONE else View.VISIBLE
+        binding.saveButton.visibility = if (editMode) View.VISIBLE else View.GONE
+    }
+
+    private fun updatePost() {
+        val updatedPost = post.value?.copy(
+            title = binding.titleEditText.text.toString(),
+            content = binding.contentEditText.text.toString(),
+            price = binding.priceEditText.text.toString().toDoubleOrNull() ?: 0.0,
+            rooms = binding.roomsEditText.text.toString().toIntOrNull() ?: 0,
+            floor = binding.floorEditText.text.toString().toIntOrNull() ?: 0,
+            updateTime = Date().time
+        )
+
+        updatedPost?.let {
+            updatePost(it)
+        }
+    }
+
+    private fun validateInput(): Boolean {
+        return binding.titleEditText.text.isNotEmpty() &&
+                binding.contentEditText.text.isNotEmpty() &&
+                binding.priceEditText.text.isNotEmpty() &&
+                binding.roomsEditText.text.isNotEmpty() &&
+                binding.floorEditText.text.isNotEmpty()
+    }
+
 }
